Added angle calculation to Line2D and moved DelaunayVoronoi
